[CPP] Make cpp client and c/cpp wrapper client tests to be consistence
