+++ aeron-client/src/test/cpp/CMakeLists.txt
@@ -37,10 +37,20 @@ endif ()
 
 function(aeron_client_test name file)
     add_executable(${name} ${file})
+    target_include_directories(${name} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
     target_link_libraries(${name} aeron_client aeron_driver aeron_client_test ${GMOCK_LIBS} ${CMAKE_THREAD_LIBS_INIT})
     target_compile_definitions(${name} PUBLIC "_SILENCE_TR1_NAMESPACE_DEPRECATION_WARNING")
     add_dependencies(${name} gmock)
     add_test(NAME ${name} COMMAND ${name})
+endfunction()
+
+function(aeron_client_wrapper_test wrapper_name file)
+    add_executable(${wrapper_name} ${file} ${TEST_HEADERS})
+    target_include_directories(${wrapper_name} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+    target_link_libraries(${wrapper_name} aeron_client_wrapper aeron_driver ${GMOCK_LIBS} ${CMAKE_THREAD_LIBS_INIT})
+    target_compile_definitions(${wrapper_name} PUBLIC "_SILENCE_TR1_NAMESPACE_DEPRECATION_WARNING" "AERON_C_CLIENT_WRAPPER_TEST")
+    add_dependencies(${wrapper_name} gmock)
+    add_test(NAME ${wrapper_name} COMMAND ${wrapper_name})
 endfunction()
 
 aeron_client_test(clientConductorTest ClientConductorTest.cpp)
@@ -66,4 +76,15 @@ aeron_client_test(errorLogReaderTest concurrent/ErrorLogReaderTest.cpp)
 aeron_client_test(oneToOneRingBufferTest concurrent/OneToOneRingBufferTest.cpp)
 aeron_client_test(channelUriStringBuilderTest ChannelUriStringBuilderTest.cpp)
 aeron_client_test(channelUriTest ChannelUriTest.cpp)
-#aeron_client_test(systemTest SystemTest.cpp)
+aeron_client_test(systemTest SystemTest.cpp)
+aeron_client_wrapper_test(systemTestW SystemTest.cpp)
+aeron_client_test(pubSubTest ../cpp_wrapper/PubSubTest.cpp)
+aeron_client_wrapper_test(pubSubTestW ../cpp_wrapper/PubSubTest.cpp)
+aeron_client_test(countersTest ../cpp_wrapper/CountersTest.cpp)
+aeron_client_wrapper_test(countersTestW ../cpp_wrapper/CountersTest.cpp)
+aeron_client_test(imagePollTest ../cpp_wrapper/ImageTest.cpp)
+aeron_client_wrapper_test(imagePollTestW ../cpp_wrapper/ImageTest.cpp)
+aeron_client_test(localAddressesTest ../cpp_wrapper/LocalAddressesTest.cpp)
+aeron_client_wrapper_test(localAddressesTestW ../cpp_wrapper/LocalAddressesTest.cpp)
+aeron_client_test(multiDestinationTest ../cpp_wrapper/MultiDestinationTest.cpp)
+aeron_client_wrapper_test(multiDestinationTestW ../cpp_wrapper/MultiDestinationTest.cpp)
