+++ aeron-client/src/test/cpp_wrapper/CountersTest.cpp
@@ -21,7 +21,7 @@
 
 #include "EmbeddedMediaDriver.h"
 #include "Aeron.h"
-#include "TestUtil.h"
+#include "util/TestUtils.h"
 
 #define COUNTER_LABEL "counter label"
 #define COUNTER_TYPE_ID (102)
