+++ aeron-client/src/test/cpp_wrapper/MultiDestinationTest.cpp
@@ -23,7 +23,7 @@
 #include "EmbeddedMediaDriver.h"
 #include "Aeron.h"
 #include "ChannelUriStringBuilder.h"
-#include "TestUtil.h"
+#include "util/TestUtils.h"
 
 using namespace aeron;
 using testing::MockFunction;
