+++ aeron-client/src/test/cpp_wrapper/PubSubTest.cpp
@@ -23,9 +23,10 @@
 #include "EmbeddedMediaDriver.h"
 #include "Aeron.h"
 #include "ChannelUriStringBuilder.h"
-#include "TestUtil.h"
+#include "util/TestUtils.h"
 
 using namespace aeron;
+using namespace aeron::test;
 using testing::MockFunction;
 using testing::_;
